fix(number): move source files to component root to fix linker errors and update schema
