+++ components/truma_inetbox/TrumaHeaterNumber.h
@@ -1,14 +1,13 @@
 #pragma once
 
-#include "enum.h"
+#include "TrumaNumberEnums.h"
 #include "esphome/components/number/number.h"
 #include "esphome/components/truma_inetbox/TrumaiNetBoxApp.h"
 
 namespace esphome {
 namespace truma_inetbox {
 
-
-class TrumaAirconManualNumber : public Component, public number::Number, public Parented<TrumaiNetBoxApp> {
+class TrumaHeaterNumber : public Component, public number::Number, public Parented<TrumaiNetBoxApp> {
  public:
   void setup() override;
   void dump_config() override;
